backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
twscrape==0.15.0
anthropic>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.7.0
